refactor: api 모듈과 domain 모듈을 연동한다
