--- app-api/src/main/kotlin/site/techmoa/app/article/domain/ArticleContent.kt
@@ -1,8 +1,0 @@
-package site.techmoa.app.article.domain
-
-import site.techmoa.app.blog.domain.Blog
-
-data class ArticleContent(
-    val article: Article,
-    val blog: Blog
-)
