--- app-api/src/main/kotlin/site/techmoa/app/core/Page.kt
@@ -1,8 +1,0 @@
-package site.techmoa.app.core
-
-data class Page<T>(
-    val data: List<T>,
-    val hasNext: Boolean,
-    val nextCursor: Long? = null,
-) {
-}
