+++ app-core/src/main/kotlin/site/techmoa/app/core/article/ArticleContent.kt
@@ -1,0 +1,8 @@
+package site.techmoa.app.core.article
+
+import site.techmoa.app.core.blog.Blog
+
+data class ArticleContent(
+    val article: Article,
+    val blog: Blog
+)
