+++ app-core/src/main/kotlin/site/techmoa/app/core/blog/BlogUseCase.kt
@@ -1,0 +1,10 @@
+package site.techmoa.app.core.blog
+
+import org.springframework.stereotype.Service
+
+@Service
+class BlogUseCase(
+    private val blogFinder: BlogFinder
+){
+
+}
