+++ app-core/src/main/kotlin/site/techmoa/app/core/common/OffsetLimit.kt
@@ -1,4 +1,4 @@
-package site.techmoa.core.common
+package site.techmoa.app.core.common
 
 data class OffsetLimit(
     val offset: Int = 0,
