+++ app-core/src/main/kotlin/site/techmoa/app/core/common/Page.kt
@@ -1,4 +1,4 @@
-package site.techmoa.core.common
+package site.techmoa.app.core.common
 
 data class Page<T>(
     val data: List<T>,
