--- app-core/src/main/kotlin/site/techmoa/core/article/ArticleContent.kt
@@ -1,8 +1,0 @@
-package site.techmoa.core.article
-
-import site.techmoa.core.blog.Blog
-
-data class ArticleContent(
-    val article: Article,
-    val blog: Blog
-)
