+++ app-storage/app-storage-db/src/main/kotlin/site/techmoa/app/storage/db/entity/BlogEntity.kt
@@ -1,7 +1,7 @@
 package site.techmoa.app.storage.db.entity
 
 import jakarta.persistence.*
-import site.techmoa.core.blog.BlogStatus
+import site.techmoa.app.core.blog.BlogStatus
 
 @Entity
 @Table(name = "blog")
